updating comments and remaving some dependencies
