+++ package.json
@@ -8,7 +8,6 @@
     "@fontsource/roboto": "^4.5.8",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.11.16",
-    "@mui/x-data-grid": "^6.5.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -16,13 +15,11 @@
     "formik": "^2.2.9",
     "i18next": "^23.1.0",
     "json-loader": "^0.5.7",
-    "mui-virtualized-grid": "^1.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "react-simple-keyboard": "^3.5.69",
-    "react-virtualized": "^9.22.5",
     "react-virtuoso": "^4.3.8",
     "react-window": "^1.8.9",
     "save": "^2.9.0",
